// src/app/resources/recibos_procesamiento/recibos-procesamiento.models.ts
export interface ReciboProcesarRequest {
    archivo: string;
    tipo: string;
}

export interface ReciboProcesarResponse {
    legajo: number;
    neto: number;
    nombreCompleto: string;
    periodo: string;
    tipo: string;
    descripcion: string;
    timeInMilis: string;
}

export interface RecibosProcesarApiResponse {
    status: string;
    data: ReciboProcesarResponse[];
}

export interface ReciboTabla extends ReciboProcesarResponse {
    aprobado: boolean;
    observaciones: string;
}

export interface ReciboAprobarRechazarRequest extends ReciboProcesarResponse {
    aprobado: boolean;
    observaciones: string;
}